@@ -10,10 +10,10 @@
         }
     ],
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.1",
 
         "doctrine/orm": "^2.4",
-        "symfony/property-access": "^4.4|^5.4|^6.0"
+        "symfony/property-access": "^5.4 || ^6.4"
     },
     "autoload": {
         "psr-4": { "SyliusLabs\\AssociationHydrator\\": "src/" }
